Move adc code into single file
